fixed data type issue when reading custom gate binary data.

+++ Online FCS Analysis/Utilities/GatePolygon.cs
@@ -17,7 +17,7 @@ namespace Online_FCS_Analysis.Utilities
     [Serializable]
     public class SPoint
     {
-        public int x { get; set; }
-        public int y { get; set; }
+        public float x { get; set; }
+        public float y { get; set; }
     }
 }
